Allow docfx to be set by an environmental variable

+++ src/VoltProjects.Server/Core/SiteCache/SiteCacheManager.cs
@@ -207,6 +207,9 @@ public sealed class SiteCacheManager
 
             void Cleanup()
             {
+                if(!_config.CleanupOnError)
+                    return;
+                
                 Directory.Delete(fullProjectDirectory, true);
             }
